refactor(chat): rename message thread-read to read
Align per-parent look endpoint with room-level read naming: POST …/messages/{messageId}/read.
